refactor: PathVariableExtractor 클래스를 web/utils 패키지로 이동
