+++ src/main/java/woowa/camp/jspcafe/web/servlet/user/UserEditServlet.java
@@ -16,7 +16,7 @@ import org.slf4j.LoggerFactory;
 import woowa.camp.jspcafe.domain.User;
 import woowa.camp.jspcafe.repository.dto.UserUpdateRequest;
 import woowa.camp.jspcafe.service.UserService;
-import woowa.camp.jspcafe.utils.PathVariableExtractor;
+import woowa.camp.jspcafe.web.utils.PathVariableExtractor;
 
 @WebServlet(name = "userUpdateServlet", value = "/users/edit/*")
 public class UserEditServlet extends HttpServlet {
